Localize MASM lint suppression policy
